Converted to conversation.ts.
